+++ VRRS_Project/assets/Icons/BarIcons.js
@@ -1,0 +1,14 @@
+import homeIcon_C from "./Bar_Icons/Home-click.png";
+import homeIcon from "./Bar_Icons/Home.png";
+import dicIcon_C from "./Bar_Icons/Dic-click.png";
+import dicIcon from "./Bar_Icons/Dic.png";
+import readIcon from "./Bar_Icons/Reading.png";
+import recommed_C from "./Bar_Icons/recommend-click.png";
+import recommend from "./Bar_Icons/recommend.png";
+import user_C from "./Bar_Icons/User-click.png";
+import user from "./Bar_Icons/User.png";
+
+export default barIcons = {
+  homeIcon: require("./Bar_Icons/Home.png"),
+  homeIcon_C: require("./Bar_Icons/Home-click.png"),
+};
